refactor: Externalize import/export functionality from Arc Core
Major architectural change to simplify Arc Core by removing import/export
functionality. Arc is now focused on being a high-performance time-series
storage and query engine only.

Breaking Changes:
- Removed InfluxDB, TimescaleDB, Kafka, and HTTP-JSON import/export endpoints
- Removed ConnectionManager in favor of simpler StorageManager
- Storage backends (MinIO/S3/GCS/Local) now managed via StorageManager only

Changes:
- Remove 11 import/export files (~2,500+ lines):
  - exporter/{__init__,http_json,influx1x,influx2x,timescale,state_manager}.py
  - api/{http_json_routes,scheduler,connection_pool,database}.py
- Replace ConnectionManager with StorageManager (storage backends only)
- Implement singleton SQLite connection pool to fix race conditions in
  multi-worker environments (fixes "database is locked" errors)
- Extract CompactionLock to api/compaction_lock.py
- Remove tombstone-based deletion code (~600 lines total):
  - storage/compaction.py: Remove _download_tombstones, _merge_tombstone_files,
    _cleanup_old_tombstones methods
  - api/duckdb_engine.py: Remove _apply_tombstone_filter, _load_tombstone_hashes,
    _compute_row_hash methods
  - ingest/arrow_writer.py: Remove _compute_row_hash, _add_row_hashes_to_records,
    _add_row_hashes_to_columns methods
- Fix compaction issues:
  - Remove ORDER BY clause causing DuckDB "invalid TType" errors with large file counts
  - Add proper exception logging in compaction cycle
  - Fix missing imports in compaction_lock.py

Impact:
- Codebase reduced from ~20K to ~17K lines (15% reduction)
- Simplified architecture focused on core time-series storage
- Improved multi-worker startup performance (no database lock contention)
- Compaction working reliably with ZSTD compression (32-34% space savings)

Migration Notes:
- Users needing import/export should use external tools or plugins
- Storage connection management now via StorageManager API
- Deletion uses rewrite-based approach (see docs/DELETE.md)

Testing:
- ✅ Multi-worker startup (42 workers) without race conditions
- ✅ Storage connection properly activated on startup
- ✅ Compaction tested: 73 files → 3 files, 20.9 MB → 14.1 MB (0.2s)
- ✅ Line protocol and MessagePack write services operational
