+++ arc.conf
@@ -44,7 +44,7 @@ worker_connections = 1000
 # ======================
 [auth]
 # Enable API token authentication (recommended: true)
-enabled = false
+enabled = true
 
 # Default API token (leave empty to auto-generate on first run)
 # IMPORTANT: Change this in production!
